Refactored conversion factory this is the most up to date submitted version

+++ app/src/main/java/com/example/myapp/Convertor/Millimeter.java
@@ -1,39 +1,47 @@
 package com.example.myapp.Convertor;
 
+// Class implementing Converts interface for Millimeter unit conversion
 public class Millimeter implements Converts {
 
+    // Method to get the name of the source unit
     @Override
     public String from() {
-        return "Milimeters";
+        return "Milimeters"; // Return the name of the source unit
     }
 
+    // Method to convert to the same unit (identity conversion)
     @Override
     public Double toSelf(Double value) {
-        return value;
+        return value; // Return the input value unchanged
     }
 
+    // Method to convert to Millimeters (identity conversion)
     @Override
     public Double toMillimeters(Double value) {
-        return toSelf(value);
+        return toSelf(value); // Return the input value unchanged
     }
 
+    // Method to convert to Centimeters
     @Override
     public Double toCentimeters(Double value) {
-        return (Double) value / 10;
+        return value / 10; // Convert to centimeters by dividing by 10
     }
 
+    // Method to convert to Meters
     @Override
     public Double toMeters(Double value) {
-        return (Double) value / 1000;
+        return value / 1000; // Convert to meters by dividing by 1000
     }
 
+    // Method to convert to Miles
     @Override
     public Double toMiles(Double value) {
-        return (Double) value * 0.000000621371f;
+        return value * 0.000000621371f; // Convert to miles using a predefined conversion factor
     }
 
+    // Method to convert to Feet
     @Override
     public Double toFeet(Double value) {
-        return (Double) value * 0.00328084f;
+        return value * 0.00328084f; // Convert to feet using a predefined conversion factor
     }
 }
